Fix getBaseRef to work when a directory has the same name as a branch (#16)
## Summary:
Now that we're using the get-changed-files action, it might make sense to refactor the use of getBaseRef 🤔 because we only need it to calculate the changed files list. But anyway, this bug is easy to fix.

Issue: FEI-3819

## Test plan:
create a branch that's the same name as a folder in this project, then make a branch off of that and to a pull-request. It shouldn't error out.
Here's the PR I made to test it https://github.com/Khan/actions-utils/pull/17

Author: jaredly

Reviewers: kevinbarabash

Required Reviewers: 

Approved by: kevinbarabash

Checks: ✅ Flow Coverage, ✅ Jest Coverage, ✅ Jest, ✅ Eslint, ✅ autofix, ✅ lint_and_unit

Pull request URL: https://github.com/Khan/actions-utils/pull/16